Exposing generateHeaderString
