+++ .eslintignore
@@ -1,1 +1,2 @@
 *.node*.js
+node_modules
